Imporove directory structure vol.1
